python-lib/pyproject.toml
[project]
name = "mahjong-detector"
version = "0.0.3"
description = "Detect mahjong tiles from image"
authors = [
    {name = "ssttkkl", email = "huang.wen.long@hotmail.com"},
]
dependencies = [
    "Pillow>=11.2.1",
    "numpy>=2.2.4",
    "onnxruntime>=1.21.0"
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "ruff<1.0.0,>=0.0.275",
    "isort<6.0.0,>=5.10.1",
    "black<24.0.0,>=23.1.0",
    "pre-commit<4.0.0,>=3.1.0",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
lint.select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
lint.ignore = ["C901", "E402", "F403", "T201"]

line-length = 160
target-version = "py38"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false



[tool.pdm]
distribution = true
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


